Day 10 2025: Part 2 was black magic using Z3 carried by GPT

+++ GlobalMethods/Globals/CreateDay.java
@@ -73,8 +73,8 @@ public class CreateDay {
 
 				    public class %s {
 				        public static void main(String[] args) {
-				            int result1 = 0;
-				            int result2 = 0;
+				            long result1 = 0;
+				            long result2 = 0;
 				            List<String> lines = ResourceLoader.getContentAsLines("%s", "Day_%s" + File.separator + "Test.txt");
